requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
courlan==1.3.2
cryptography==46.0.3
dataclasses-json==0.6.7
dateparser==1.2.2
distro==1.9.0
docx2txt==0.9
emoji==2.15.0
et_xmlfile==2.0.0
faiss-cpu==1.13.1
fastapi==0.124.4
filetype==1.2.0
frozenlist==1.8.0
h11==0.16.0
html5lib==1.1
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.2
langchain-ollama==1.0.1
langchain-openai==1.1.4
langchain-text-splitters==1.1.0
langdetect==1.0.9
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
lxml==6.0.2
lxml_html_clean==0.4.3
markdownify==1.2.2
marshmallow==3.26.1
msoffcrypto-tool==6.0.0
multidict==6.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
nltk==3.9.2
numpy==2.3.5
olefile==0.47
ollama==0.6.1
openai==2.13.0
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pip==25.3
propcache==0.4.1
psutil==7.2.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
PyMuPDF==1.26.7
pymupdf4llm==0.2.9
PyMySQL==1.1.2
pypdf==6.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.21
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
unstructured==0.18.27
unstructured-client==0.42.8
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.38.0
webencodings==0.5.1
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
chromadb
langchain-chroma
